backend/requirements.txt
flask
flask-cors
gunicorn
tensorflow-cpu==2.13.0
h5py
opencv-python-headless
pillow
numpy
reportlab
gdown